Added all structs to structs.h
