+++ inc/cmd_list.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cmd_list.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pipolint <pipolint@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ehammoud <ehammoud@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/22 05:47:46 by marvin            #+#    #+#             */
-/*   Updated: 2024/07/14 16:34:15 by pipolint         ###   ########.fr       */
+/*   Updated: 2024/07/16 13:03:20 by ehammoud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,29 +15,7 @@
 
 # include "queues.h"
 # include "env_list.h"
-
-typedef struct s_cmd
-{
-	int				last_fd;
-	int				status;
-	int				in_fd;
-	int				out_fd;
-	int				params_cnt;
-	int				infile_cnt;
-	int				outfile_cnt;
-	char			rep;
-	char			**infiles;
-	int				*in_flags;
-	int				*out_flags;
-	char			**outfiles;
-	char			**params;
-	t_bool			heredoc;
-	t_bool			or_op;
-	int				before;
-	int				after;
-	t_queue			**parse;
-	struct s_cmd	*next;
-}	t_cmd;
+# include "general.h"
 
 t_cmd	*new_cmd_node(char **params, t_queue **q);
 void	add_cmd_node(t_cmd **cmd, t_cmd *to_add);
